Restructure library, fix license
This patch restructures the library to remove the `third_party/`
subdirectory. In addition it updates the `Cargo.toml` to express that it
is also ISC licensed.

Finally, it also bumps the version.
