+++ tests/assets/json5_dot_org_example.json5
